pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forgeoagent"
version = "0.1.0"
authors = [
    {name = "Angel Koradiya", email = "Koradiyaangel11@gmail.com"},
]
description = "ForgeOAgent - An AI agent framework"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.25.0",
    "google-api-core>=2.25.0",
    "google-generativeai>=0.8.0",
    "fastapi>=0.119.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
    "jinja2>=3.1.0",
]

[project.urls]
Homepage = "https://forgeoagent.vercel.app"
Documentation = "https://github.com/angelkoradiya/forgeoagent#readme"
Repository = "https://github.com/angelkoradiya/forgeoagent"
Issues = "https://github.com/angelkoradiya/forgeoagent/issues"

[project.scripts]
forgeoagent = "forgeoagent.cli:main"

[tool.setuptools]
packages = ["forgeoagent"]

[tool.setuptools.package-data]
forgeoagent = ["web/templates/*", "web/static/*"]